+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- Added widget details for Textual containers and base classes (https://github.com/TomJGooding/textual-dissect/pull/13).
+- Selecting a widget in the 'Inheritance Tree' will navigate to that widget (https://github.com/TomJGooding/textual-dissect/pull/13).
+
+### Changed
+
+- Changed the often redundant 'Module Widgets' list to 'Child Widgets' (https://github.com/TomJGooding/textual-dissect/pull/13).
+
 ## [0.1.0] - 2025-06-22
 
 - Initial release